feat(types): Add subCommands command option
This mainly for use in settings  #22 but could be useful in other ways.

The option just tells the message handler to treat the first argument
of the command as a sub command.

this is only available in the message handler not the socket handler,
because the subcommands are normal commands and can be called
without the parent.
So this is just for user interface purposes.